refactor: move more router stuff to the router lib
